Changed Chimera.js and arms/torsos
